backend/requirements.txt
fastapi[standard]
uvicorn
tensorflow
pillow
opencv-python
numpy
python-multipart